requirements.txt
alembic==1.10.2
aniso8601==9.0.1
asgiref==3.6.0
autopep8==2.0.1
Babel==2.11.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
crispy-bootstrap5==0.7
Django==4.1.7
django-better-admin-arrayfield==1.4.2
django-ckeditor==6.5.1
django-cleanup==7.0.0
django-crispy-forms==2.0
django-enum==1.1.2
django-filter==22.1
django-js-asset==2.0.0
django-jsonform==2.15.0
django-money==3.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
Pillow==9.4.0
psycopg2==2.9.5
py-moneyed==2.0
pycodestyle==2.10.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
SQLAlchemy-ImageAttach==1.1.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
Wand==0.6.11
Werkzeug==2.2.3
